PCI: keystone: remove the invoke of serdes phy init
Since SerDes PHY provider now performs the common
portion of SerDes configurations (configurations
that do not depend on whether a lane will actually
be used) and that is all what PCIe needs as far as
SerDes configurations are concerned.  Hence there
is no real need for PCIe driver to receive the PHY
instance.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>
Signed-off-by: Jacob Stiffler <j-stiffler@ti.com>